Fixed admin config section ordering bug
The config section sort_order was updated to respect the containing
tab's sort order in addition to the section sort order.  Now sections
are sorted in the context of the tab and the final list of sections is
built first on the tab sort_order, then the section sort_order.  This
fixes a problem with sections that do not have a sort_order becoming the
default section.  The default section is now the first section of the
first tab.