Fixing broken classname resolver during init stage
During init stage, when the config XML is not fully loaded, if certain
tasks fail, they throw exceptions and use the core helper to translate
the messages.

If the classname resolver cannot find a specific config section for a
requested module, it defaults to treating the module alias as a
sub-namespace under Mage.  So, a request for 'core/data' would, in the
absence of a valid config, resolve the module portion to Mage_Core.

This was fine until a check for the deprecatedNode element was added in
1.6 that broke this process.  The check is assuming that the $config
variable is a valid config section when it could actually be a NULL
instead. Since the check ignores this possible outcome, it tries to
access a property on a NULL value causing a PHP error.

This patch simply wraps the vulnerable code section and checks that
$config is not a null value.