Fixing URL generation for admin and in APIs
In the admin area (and in api.php) the URL generator was forcing the
usage of index.php even if the use of rewrites was enabled.  This
resulted in URLs for admin always starting with /index.php/admin/ and
looking ugly.

I removed the check for being in the admin area and re-ordered the two
remaining checks (installed, rewrites enabled) to optimize their order.