Fix minor bug in secret key verification
When using an internal forward, the module/controller/action changes,
but the secret key does not.  This could cause problems with the
internal getSecretKey() method on the Mage_Adminhtml_Model_Url class.
In an effort to prevent this, the Magento Core team implemented a manual
parsing of the original PATH_INFO on the request to pull out the
controller/action names.  They also default back to using the controller
and/or action stored in the request if the path parsing returns nothing.

This is a flawed procedure as it's dependant upon knowing how the router
is parsing a URL into a module/controller/action and then performing the
same action.  With the current system routers, this happens to work.  If
you change routers, it will likely fail. Fortunatly the request object
has a method getBeforeForwardInfo() that is populated when you call
initForward() on the request.  The base controller classes in Magento
call initForward() inside of the _forward() method.  Giving this
situation, you can us getBeforeForwardInfo() to extract the original
module/controller/action of the request to verify the secret key.