Fix bad usage of 'false' as a string instead of a boolean
The initial config for the tree node root element used in role resource editing
was broken.  The 'checked' attribute was being set to a string of the word 'false'
when it should have been the false boolean literal.  This was causing the root
node of the tree to be included in the selected resource list.  The bug is minor
and had no serious side-effects, but was still a bug that caused __root__ to be
sent as a selected resource.