Add Australian regions to the default list.
The regions have been added to the data-install-1.6.0.0.php file to
ensure better forward compatibility with future versions of Magento,
which may create new versions of the Mage_Directory module, along with
new data-upgrade files, which will overwrite anything we put in here and
require a merge. This makes future upstream merges easier at the expense
of existing websites not automatically getting the changes as part of an
upgrade.

If you are already running an existing website you can add these regions
by installing the Fontis_Australia extension, or adding the regions
individually.